[TG-22] feat: validacio automatica Creacio de la logica per fer la validacio automatica d'una incidencia per a poder-la tancar. S'ha afegit al calcul del vot la reputacio de l'usuari. Al crear o eliminar un vot es calcula si ja es pot tancar l'incidencia utilitzant el THRESHOLD indicant quina puntuacio ha d'arribar la "reliability" d'una incidencia. Pasat la reputacio a double.

+++ src/main/java/com/safesteps/backend/domain/users/model/User.java
@@ -42,7 +42,7 @@ public class User {
     private Boolean isAnonymous = false;
 
     @Column(nullable = false)
-    private Integer reputacio = 1;
+    private double reputacio = 1;
 
     @CreationTimestamp
     @Column(name = "created_at", updatable = false)
